# Documentacion/openapi.yaml
openapi: 3.0.1
info:
  title: API_RPG_DOCUMENTACION
  description: Documentacion
  contact:
    name: Documentacion de puntos de acceso a la API
    url: url
    email: laureano.de.sousa@iessanmamede.com
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/tipoGrupo/{id}:
    get:
      tags:
        - TipoGrupo
      summary: Obtener un tipo de grupo por ID
      operationId: obtener
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
        - TipoGrupo
      summary: Actualizar un tipo de grupo por ID
      operationId: actualizar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoGrupo'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
        - TipoGrupo
      summary: Eliminar un tipo de grupo por ID
      operationId: borrar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/efectos/{id}:
    get:
      tags:
        - EfectoEstado
      summary: Obtener un efecto de estado por ID
      operationId: obtener_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
        - EfectoEstado
      summary: Actualizar un efecto de estado por ID
      operationId: actualizar_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EfectoEstado'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
        - EfectoEstado
      summary: Eliminar un efecto de estado por ID
      operationId: borrar_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/tipoGrupo:
    post:
      tags:
        - TipoGrupo
      summary: Crear un nuevo tipo de grupo
      operationId: guardar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoGrupo'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/efectos:
    post:
      tags:
        - EfectoEstado
      summary: Crear un nuevo efecto de estado
      operationId: guardar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EfectoEstado'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    TipoGrupo:
      type: object
      properties:
        tipo_grupo_id:
          type: integer
          description: ID único del tipo de grupo.
          format: int64
          example: 1
        nombre:
          type: string
          description: Nombre del tipo de grupo.
          example: Equipo de Aventureros
        numeroIntegrantesMax:
          type: integer
          description: Número máximo de integrantes permitidos en el grupo.
          format: int32
          example: 5
        compartenExpDrops:
          type: boolean
          description: Indica si los miembros del grupo comparten experiencia y drops.
          example: true
        descripcion:
          type: string
          description: Descripción del tipo de grupo.
          example: Grupo para misiones de alto nivel.
      description: Entidad que representa un tipo de grupo en el sistema.
...
    EstadisticasGenerales:
      type: object
      properties:
        estadistica_id:
          type: integer
          description: ID único de las estadísticas generales.
          format: int64
          example: 1
        vida:
          type: integer
          description: Vida base de la entidad.
          format: int32
          example: 100
        regeneracionVida:
          type: integer
          description: Regeneración de vida base por intervalo de tiempo.
          format: int32
          example: 5
        escudo:
          type: integer
          description: Escudo base de la entidad.
          format: int32
          example: 50
        energia:
          type: integer
          description: Energía base de la entidad.
          format: int32
          example: 100
        regeneracionEnergia:
          type: integer
          description: Regeneración de energía base por intervalo de tiempo.
          format: int32
          example: 10
        mana:
          type: integer
          description: Maná base entidad.
          format: int32
          example: 100
        regeneracionMana:
          type: integer
          description: Regeneración de maná base por intervalo de tiempo.
          format: int32
          example: 5
        ataqueFisico:
          type: integer
          description: Ataque físico base entidad.
          format: int32
          example: 20
        ataqueMagico:
          type: integer
          description: Ataque mágico base entidad
          format: int32
          example: 15
        defensaFisica:
          type: integer
          description: Defensa física entidad.
          format: int32
          example: 10
        defensaMagica:
          type: integer
          description: Defensa mágica entidad.
          format: int32
          example: 10
      description: Modelo que representa las estadísticas generales de un personaje, monstruo o efecto.
...
    EfectoEstado:
      type: object
      properties:
        efecto_id:
          type: integer
          description: ID único del efecto
          format: int64
          example: 1
        imagen:
          type: string
          description: URL del icono del efecto
          example: potenciacion.png
        nombre:
          type: string
          description: Nombre del efecto
          example: Fuerza
        descripcion:
          type: string
          description: Descripción del efecto
          example: Aumenta el ataque en 5 puntos
        tipo:
          type: string
          description: Tipo de efecto (buff, debuff)
          example: buff
          enum:
            - BUFF
            - DEBUFF
        tipo_afectado:
          type: string
          description: Tipo de entidad afectada (personaje, monstruo, todo)
          example: personaje
          enum:
            - PERSONAJE
            - MONSTRUO
            - TODO
        duracion_efecto:
          type: number
          description: Duración del efecto en turnos
          format: double
          example: 5
        intervalos_efecto:
          type: number
          description: Intervalos de aplicación del efecto
          format: double
          example: 1
        acumulaciones:
          type: integer
          description: Número de acumulaciones del efecto
          format: int32
          example: 0
        estadisticas:
          $ref: '#/components/schemas/EstadisticasGenerales'
      description: Entidad que representa un efecto o estado en el sistema
...